// apps/web/src/api/generated/RunDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type RunDto = { id: string, workflow_id: string, workflow_version: string, status: string, started_at: string | null, completed_at: string | null, error: string | null, created_at: string, run_label: string | null, execution_profile: string | null, predecessor_run_id: string | null, };