// apps/web/src/api/generated/ToolDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ToolDto = { id: string, kind: string, target_id: string, display_name: string, description: string | null, category: string, tags: Array<string>, icon: string | null, extension_id: string, availability: string, };